requirements.txt
tensorflow-gpu==2.8.0
matplotlib
soundfile
webrtcvad
jupyterlab
notebook
pyloudnorm
librosa